fix: Move `virtio_gen` into module under `vmm`
