+++ src/vmm/src/devices/virtio/block/device.rs
@@ -20,10 +20,10 @@ use crate::rate_limiter::{BucketUpdate, RateLimiter};
 use serde::{Deserialize, Serialize};
 use utils::eventfd::EventFd;
 use utils::kernel_version::{min_kernel_version_for_io_uring, KernelVersion};
-use virtio_gen::virtio_blk::{
+use crate::virtio_gen::virtio_blk::{
     VIRTIO_BLK_F_FLUSH, VIRTIO_BLK_F_RO, VIRTIO_BLK_ID_BYTES, VIRTIO_F_VERSION_1,
 };
-use virtio_gen::virtio_ring::VIRTIO_RING_F_EVENT_IDX;
+use crate::virtio_gen::virtio_ring::VIRTIO_RING_F_EVENT_IDX;
 use vm_memory::GuestMemoryMmap;
 
 use super::super::{ActivateResult, DeviceState, Queue, VirtioDevice, TYPE_BLOCK};
