+++ src/vmm/src/devices/virtio/block/event_handler.rs
@@ -103,7 +103,7 @@ pub mod tests {
     use std::sync::{Arc, Mutex};
 
     use event_manager::{EventManager, SubscriberOps};
-    use virtio_gen::virtio_blk::{VIRTIO_BLK_S_OK, VIRTIO_BLK_T_OUT};
+    use crate::virtio_gen::virtio_blk::{VIRTIO_BLK_S_OK, VIRTIO_BLK_T_OUT};
     use vm_memory::{Bytes, GuestAddress};
 
     use super::*;
