+++ src/vmm/src/devices/virtio/persist.rs
@@ -10,7 +10,7 @@ use std::sync::{Arc, Mutex};
 use snapshot::Persist;
 use versionize::{VersionMap, Versionize, VersionizeResult};
 use versionize_derive::Versionize;
-use virtio_gen::virtio_ring::VIRTIO_RING_F_EVENT_IDX;
+use crate::virtio_gen::virtio_ring::VIRTIO_RING_F_EVENT_IDX;
 use vm_memory::address::Address;
 use vm_memory::{GuestAddress, GuestMemoryMmap};
