+++ src/vmm/src/virtio_gen/virtio_ring.rs
